Fixed compiler warning about unsafe type-conversion.

+++ src/libxsmm_bgemm.c
@@ -141,7 +141,7 @@ LIBXSMM_API_DEFINITION libxsmm_bgemm_handle* libxsmm_bgemm_handle_create(
             /* TODO: more sophisticated strategy perhaps according to CPUID */
             const char *const env_p = getenv("LIBXSMM_BGEMM_PREFETCH");
             const int uid = ((0 == env_p || 0 == *env_p) ? 7/*LIBXSMM_PREFETCH_AL2BL2_VIA_C*/ : atoi(env_p));
-            descriptor.prefetch = libxsmm_gemm_uid2prefetch(uid);
+            descriptor.prefetch = (unsigned short)libxsmm_gemm_uid2prefetch(uid);
           }
           else { /* user-defined */
             descriptor.prefetch = (unsigned short)prefetch;
